Swapped from sysfs_gpio to rppal, which is lower latency and supports i2c

+++ alarm_clock/src/config.rs
@@ -2,7 +2,7 @@ use serde::Deserialize;
 
 #[derive(Deserialize)]
 pub struct Config {
-	pub buzzer_pin: u64,
-	pub button_pins: Vec<u64>,
+	pub buzzer_pin: u8,
+	pub button_pins: Vec<u8>,
 	pub midi_dir: String,
 }
